+++ CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - update `scheduled` attribute on any item state
 
+### Changed
+- Edition 2018: wip
+- up to date actix/tokio/futures/bytes and others
+- http server starts default workers count
+- clippy and rustfmt fixes
+
 ## [0.1.0] - 2020-01-02
 
 ### Added